Decouple Gene ID conversion and prep.genesets
There are two approaches to converting gene Ids. One approach converts
gene IDs of gene set collections to match those of expression profile;
the other approach converts gene IDs of expression profile to match
those of gene set collection. We suggest using the latter approach
since it avoids expanding the gene set sizes through one-to-many
mappings.

To leave this choice to users, we remove the functionality coupled with
the prep.genesets for ID conversion. Consequently, tests and dataset
are also updated to make this change.