Add a large artificial expression profile
The profile includes 20514 gene and 40 samples. The Entrez id has
been used as gene ids; the expression values are simulated by a
random normal distribution with a mean of 0 and a standard
deviation of 1. The expected behavior for most gene set analysis
method when analyzing this dataset is to not report any differentially
enriched genes.